refactor: `@/shared` folder should only ever inherit from `@/shared` to avoid bundle pollution
