+++ src/shared/model/Configuration.ts
@@ -1,0 +1,1 @@
+export type Stage = 'DEV' | 'CODE' | 'PROD';
